Remove redundant action items. (#9)
This pull request includes a change to the .github/workflows/pypi-publish.yml file to modify the conditions under which the workflow is triggered.

Workflow trigger modifications:

Removed the paths condition to trigger the workflow on changes to the package code and metadata. This means the workflow will now only trigger on version tags and pull requests to the main branch. (.github/workflows/pypi-publish.yml)